SHOP-47: FixDefaultAddress

+++ Shoppe/Modules/Profile/AddressesViewController.swift
@@ -6,11 +6,12 @@
 //
 
 import UIKit
+import FirebaseAuth
 
 class AddressesViewController: UIViewController {
     var addresses: [AddressModel] = []
     let tableView = UITableView()
-    let userId = "currentUserId"
+    var userId: String?
     var onAddressSelected: ((AddressModel) -> Void)?
     
     override func viewDidLoad() {
@@ -19,7 +20,13 @@ class AddressesViewController: UIViewController {
         view.backgroundColor = .systemBackground
         setupTableView()
         setupAddButton()
-        fetchAddresses()
+        
+        if let user = Auth.auth().currentUser {
+            userId = user.uid
+            fetchAddresses()
+        } else {
+            print("Пользователь не авторизован")
+        }
     }
     
     private func setupTableView() {
@@ -45,6 +52,8 @@ class AddressesViewController: UIViewController {
     }
 
     private func fetchAddresses() {
+        guard let userId = userId else { return }
+        
         AddressManager.shared.fetchAddresses(for: userId) { [weak self] addresses in
             self?.addresses = addresses
             DispatchQueue.main.async {
@@ -68,6 +77,7 @@ extension AddressesViewController: UITableViewDataSource, UITableViewDelegate {
     }
 
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        guard let userId = userId else { return }
         let selectedAddress = addresses[indexPath.row]
         guard selectedAddress.isDefault == false else {
             dismiss(animated: true) { [weak self] in
@@ -93,11 +103,14 @@ extension AddressesViewController: UITableViewDataSource, UITableViewDelegate {
             self?.dismiss(animated: true) { [weak self] in
                 self?.onAddressSelected?(selectedAddress) 
             }
+            
+            NotificationCenter.default.post(name: .addressUpdated, object: nil)
         }
     }
 
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         guard editingStyle == .delete else { return }
+        guard let userId = userId else { return }
         let address = addresses[indexPath.row]
         AddressManager.shared.deleteAddress(address.id, for: userId) { [weak self] success in
             if success {
